+++ crates/uv-distribution/src/source/mod.rs
@@ -1552,58 +1552,84 @@ impl<'a, T: BuildContext> SourceDistributionBuilder<'a, T> {
             return Err(Error::HashesNotSupportedGit(source.to_string()));
         }
 
-        // If this is GitHub URL, attempt to resolve to a precise commit using the GitHub API.
-        match self
-            .build_context
-            .git()
-            .github_fast_path(
-                resource.git,
-                client
-                    .unmanaged
-                    .uncached_client(resource.git.repository())
-                    .clone(),
-            )
-            .await
+        // If the reference appears to be a commit, and we've already checked it out, avoid taking
+        // the GitHub fast path.
+        let cache_shard = resource
+            .git
+            .reference()
+            .as_str()
+            .and_then(|reference| GitOid::from_str(reference).ok())
+            .map(|oid| {
+                self.build_context.cache().shard(
+                    CacheBucket::SourceDistributions,
+                    WheelCache::Git(resource.url, oid.as_short_str()).root(),
+                )
+            });
+        if cache_shard
+            .as_ref()
+            .is_some_and(|cache_shard| cache_shard.is_dir())
         {
-            Ok(Some(precise)) => {
-                // There's no need to check the cache, since we can't use cached metadata if there are
-                // sources, and we can't know if there are sources without fetching the
-                // `pyproject.toml`.
-                //
-                // For the same reason, there's no need to write to the cache, since we won't be able to
-                // use it on subsequent runs.
-                match self
-                    .github_metadata(precise, source, resource, client)
-                    .await
-                {
-                    Ok(Some(metadata)) => {
-                        // Validate the metadata, but ignore it if the metadata doesn't match.
-                        match validate_metadata(source, &metadata) {
-                            Ok(()) => {
-                                debug!("Found static metadata via GitHub fast path for: {source}");
-                                return Ok(ArchiveMetadata {
-                                    metadata: Metadata::from_metadata23(metadata),
-                                    hashes: HashDigests::empty(),
-                                });
-                            }
-                            Err(err) => {
-                                debug!("Ignoring `pyproject.toml` from GitHub for {source}: {err}");
+            debug!("Skipping GitHub fast path for: {source} (shard exists)");
+        } else {
+            debug!("Attempting GitHub fast path for: {source}");
+
+            // If this is GitHub URL, attempt to resolve to a precise commit using the GitHub API.
+            match self
+                .build_context
+                .git()
+                .github_fast_path(
+                    resource.git,
+                    client
+                        .unmanaged
+                        .uncached_client(resource.git.repository())
+                        .clone(),
+                )
+                .await
+            {
+                Ok(Some(precise)) => {
+                    // There's no need to check the cache, since we can't use cached metadata if there are
+                    // sources, and we can't know if there are sources without fetching the
+                    // `pyproject.toml`.
+                    //
+                    // For the same reason, there's no need to write to the cache, since we won't be able to
+                    // use it on subsequent runs.
+                    match self
+                        .github_metadata(precise, source, resource, client)
+                        .await
+                    {
+                        Ok(Some(metadata)) => {
+                            // Validate the metadata, but ignore it if the metadata doesn't match.
+                            match validate_metadata(source, &metadata) {
+                                Ok(()) => {
+                                    debug!(
+                                        "Found static metadata via GitHub fast path for: {source}"
+                                    );
+                                    return Ok(ArchiveMetadata {
+                                        metadata: Metadata::from_metadata23(metadata),
+                                        hashes: HashDigests::empty(),
+                                    });
+                                }
+                                Err(err) => {
+                                    debug!(
+                                        "Ignoring `pyproject.toml` from GitHub for {source}: {err}"
+                                    );
+                                }
                             }
                         }
-                    }
-                    Ok(None) => {
-                        // Nothing to do.
-                    }
-                    Err(err) => {
-                        debug!("Failed to fetch `pyproject.toml` via GitHub fast path for: {source} ({err})");
+                        Ok(None) => {
+                            // Nothing to do.
+                        }
+                        Err(err) => {
+                            debug!("Failed to fetch `pyproject.toml` via GitHub fast path for: {source} ({err})");
+                        }
                     }
                 }
-            }
-            Ok(None) => {
-                // Nothing to do.
-            }
-            Err(err) => {
-                debug!("Failed to resolve commit via GitHub fast path for: {source} ({err})");
+                Ok(None) => {
+                    // Nothing to do.
+                }
+                Err(err) => {
+                    debug!("Failed to resolve commit via GitHub fast path for: {source} ({err})");
+                }
             }
         }