+++ crates/uv-static/src/env_vars.rs
@@ -158,6 +158,11 @@ impl EnvVars {
     /// will compile Python source files to bytecode after installation.
     pub const UV_COMPILE_BYTECODE: &'static str = "UV_COMPILE_BYTECODE";
 
+    /// Equivalent to the `--no-editable` command-line argument. If set, uv
+    /// installs any editable dependencies, including the project and any workspace members, as
+    /// non-editable
+    pub const UV_NO_EDITABLE: &'static str = "UV_NO_EDITABLE";
+
     /// Equivalent to the `--no-binary` command-line argument. If set, uv will install
     /// all packages from source. The resolver will still use pre-built wheels to
     /// extract package metadata, if available.
@@ -256,6 +261,14 @@ impl EnvVars {
 
     /// Specifies the directory for storing managed Python installations.
     pub const UV_PYTHON_INSTALL_DIR: &'static str = "UV_PYTHON_INSTALL_DIR";
+
+    /// Managed Python installations information is hardcoded in the `uv` binary.
+    ///
+    /// This variable can be set to a URL pointing to JSON to use as a list for Python installations.
+    /// This will allow for setting each property of the Python installation, mostly the url part for offline mirror.
+    ///
+    /// Note that currently, only local paths are supported.
+    pub const UV_PYTHON_DOWNLOADS_JSON_URL: &'static str = "UV_PYTHON_DOWNLOADS_JSON_URL";
 
     /// Managed Python installations are downloaded from the Astral
     /// [`python-build-standalone`](https://github.com/astral-sh/python-build-standalone) project.