+++ crates/uv/src/commands/python/list.rs
@@ -102,6 +102,7 @@ pub(crate) async fn list(
         let downloads = download_request
             .as_ref()
             .map(PythonDownloadRequest::iter_downloads)
+            .transpose()?
             .into_iter()
             .flatten();